fix: wire artifact uploads through ConnectorRunner.invoke_capability
The previous commit left connector_client as None (never populated),
silently breaking uploads. Now PickConnector holds an Arc<ConnectorRunner>
and calls runner.invoke_capability() which provides full request/response
with failure confirmation routed through the SDK's internal client.

Removes dead ConnectorClient import and stale _pub wrappers.

+++ crates/tools/src/webwright/mod.rs
@@ -643,7 +643,7 @@ async fn try_sidecar_execution(
                     stderr: String::new(),
                     exit_code: 0,
                     timed_out: false, // we handled it gracefully, not a hard timeout
-                    duration_ms: (timeout_secs * 1000) as u64,
+                    duration_ms: timeout_secs * 1000,
                 });
             }
         };
